+++ CHANGELOG.md
@@ -10,9 +10,9 @@ Neste arquivo (CHANGELOG.MD) você encontrará somente as mudanças referentes a
 ### Sobre os "BREAKING CHANGES"
 Podemos ter pequenas breaking changes sem alterar o `major` version, apesar de serem pequenas, podem alterar o comportamento da funcionalidade caso não seja feita uma atualização, **preste muita atenção** nas breaking changes dentro das versões quando existirem.
 
-## [Unreleased]
+## Não publicado
 ### Adicionado
-- `QasFormGenerator`: criado slot `after-[nome-do-fieldset]`.
+- `QasFormGenerator`: criado slot `legend-bottom-[nome-do-fieldset]`.
 
 ## [3.17.0-beta.6] - 10-09-2024
 ### Corrigido